# src/swagger/swagger-config.yaml
openapi: 3.0.3
info:
  title: API de la Libreria ICPF
  description: |-
    Documentación de la API de la librería ICPF que proporciona endpoints para la gestión de libros, autores, categorias, pedidos y estadísticas.
    
    Incluye especificaciones detalladas para autenticación, operaciones CRUD y consultas avanzadas.
    
  contact:
    email: jonathancanofreta@gmail.com
  license:
    name: Licencia MIT 
    url: https://opensource.org/license/mit
  version: 1.0.0
servers:
  - url: https://icpf-api-production.up.railway.app/
    description: Servidor Railway
...
paths:
  /libros:
    get:
      tags:
        - libros
      summary: Obtener todos los libros
      description: Con el metodo GET /libros puedes obtener todos los libros disponibles en la librería
      operationId: getLibros
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Libros'          
        '400':
          description: Solicitud incorrecta
      #security:
        #- petstore_auth:
            #- write:pets
            #- read:pets
    post:
      tags:
        - libros
      summary: Add a new pet to the store
      description: Add a new pet to the store
      operationId: addPet
      requestBody:
        description: Create a new pet in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
          application/xml:
            schema:
              $ref: '#/components/schemas/Pet'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Pet'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
      security:
        - petstore_auth:
            - write:pets
            - read:pets
...
components:
  schemas:
    Libros:
      type: object
      properties:
        libro_id:
          type: integer
          format: int64
          example: 10
        libro_titulo:
          type: string
          example: Panorama del Antiguo Testamento
        libro_descripcion:
          type: string
          example: Este es un libro que recorre teológica e historicamente el relato del Antiguo Testamento...
        libro_precio:
          type: integer
          format: int64
          example: 25.34
        libro_cantidad:
          type: integer
          example: 4
        libro_paginas:
          type: integer
          format: int64
          example: 465
        libro_imagen:
          type: string
          example: https://iguiskudllwluojaxepj.supabase.co/storage/v1/object/public/icpf-imagenes/libros/...
      xml:
        name: libros
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Dogs
      xml:
        name: category
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: tag
    Pet:
      required:
        - name
        - photoUrls
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: doggie
        category:
          $ref: '#/components/schemas/Category'
        photoUrls:
          type: array
          xml:
            wrapped: true
          items:
            type: string
            xml:
              name: photoUrl
        tags:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Tag'
        status:
          type: string
          description: pet status in the store
          enum:
            - available
            - pending
            - sold
      xml:
        name: pet
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header